Fix inherited eyeball handling for nodes
Introduce the concept of logical and effective global visibility. The
slideplayer can this way check the logical global visibility (which is
based on the eyeball and nothing else) while others can continue using
the effective visibility (that also includes the visibilityTag).

Complete with a simple test scene that allows interactively toggling the
eyeball for some model nodes.

Task-number: QT3DS-1985
Change-Id: I3492412d0c80c2d2ab6e419fa6aefd035171508b
Reviewed-by: Christian Stromme <christian.stromme@qt.io>