Differentiate between two types of "dirty" lights
Changing the visibility, adding, removing a light needs rebuilding
default material shaders. Other changes to a light are fine with the
existing shaders, only the uniforms and uniform buffers have to be
updated. Make this distinction clear, in order to avoid unnecessarily
rebuilding materials in scenes where a light is part of a hierarchy that
has its transform animated for example.

Task-number: QT3DS-1989
Change-Id: I2b805593a12f3f05e0fa90b8073b5745084f1287
Reviewed-by: Andy Nichols <andy.nichols@qt.io>