Remove scene and slide callbacks when removing objects
When associating an object with a scene, or a slide with a master slide,
the scene/master slide will start listening to property and other
changes on that object. When disconnecting from the scene/master slide
these callbacks should not be active anymore. (while unlikely,
technically we could still reuse the same object in another scene or
slide graph)

To implement this, some non-node types need a bit of reshuffling when it
comes to how and when the attached object is created since we now need
to store some callback registration data in there.

Change-Id: I986863f7802b0b53a560af24eccb1dcb33958e2e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>