Do not leak attached objects when removing and readding
The attached object should go away when an object gets disconnected
from a scene (or master slide). It is recreated when the object gets
added into a scene (or master slide) hierarchy again.

To exercise this, the console's setparent() now works as expected,
supporting both unparenting (with a 'null' parent) and reparenting.

Change-Id: I70d69300d7fbab6a48b33d7082b2fb9fa9fe1e7e
Reviewed-by: Andy Nichols <andy.nichols@qt.io>