Start profileui with dpr-specific scale also in QWindow mode
The default widget mode causes the viewer to manage profileui activation
via widgets (menus, keyboard shortcuts). In QWindow mode however the
Q3DSEngine's built-in keyboard/mouse handling is used for this. Add
support for picking up the dpr-specific scale factor also for the
latter.

This applies when launching q3dsviewer with the -w argument. The
explorer tool is still not working ideally since the dpr is always 1
there for some reason.

Change-Id: I118cd4aa6e67c68513df2ee95663221d7b48fd4b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>