Make Q3DSEngine::setPresentation a first class API
Previously this was plagued by todos since it did not support
any types of subpresentations properly. However, that would not be
useful anyway: a programatically constructed and managed scene will have
no use of the "static" subpresentation concept, especially if QML is
used which provides the standard component and object model.

So drop the QVector and make the function take a single presentation
only.

Change-Id: I4bc6787dd057bfd08a3f542fa8bbac6efde774e0
Reviewed-by: Andy Nichols <andy.nichols@qt.io>