Disable dialogs while profileui is up
Parsing something, like an effect due to setting the 'class' property to
a file, can lead to parse errors that pop up dialogs by default when
running with the widget-based viewer. While not an issue for
applications or the QWindow-mode viewer, it is nice if we avoid popping
up dialogs when triggering errors from the console - these should rather
print on the debug output and show up in the log.

In addition, popping up a dialog with the console active leads to
strange problems with input afterwards. Avoid this altogether.

Change-Id: Ica5aa893601418d569b4b0dcbc26d042e6af9284
Reviewed-by: Andy Nichols <andy.nichols@qt.io>