Add support for changing the custom material source on the fly
Like with effects, the 'class' property can now be changed at run time,
and can also refer to a file (in addition to a #id from the .uip's Classes
section)

Note that constructing materials on the fly needs to be done so that the
active scene's tree is consistent at any time. For example, to replace a
model's material, the model has to first be unparented from the scene's
object tree (and so be disconnected from the scene) because having a
model without the sufficient number of materials is not allowed. Once
the child object (the material) is replaced (and, for custom materials,
the class property was set), the model can be parented back to where it was.
When just changing the class property on a CustomMaterial, this is not needed
of course, the change is picked up right away.

Change-Id: I086154023513d13a4a298aeac7e0e247951a6e09
Reviewed-by: Andy Nichols <andy.nichols@qt.io>